refactor: organize imports in falcon init
