+++ falcon/hooks.py
@@ -19,7 +19,7 @@ from inspect import getmembers
 from inspect import iscoroutinefunction
 import re
 
-from falcon import COMBINED_METHODS
+from falcon.constants import COMBINED_METHODS
 from falcon.util.misc import get_argnames
 from falcon.util.sync import _wrap_non_coroutine_unsafe
